all file split to components/ add scroll css/ fix homepage state not initial
