+++ src/index.css
@@ -3,5 +3,5 @@
 @tailwind utilities;
 
 * {
-  scroll-margin-top: 5rem;
+  scroll-margin-top: 10rem;
 }
